+++ src/main/java/ai/mcpdirect/backend/dao/mapper/aitool/ToolMakerTemplateMapper.java
@@ -1,7 +1,6 @@
 package ai.mcpdirect.backend.dao.mapper.aitool;
 
 import ai.mcpdirect.backend.dao.entity.aitool.AIPortToolMakerTemplate;
-import ai.mcpdirect.backend.dao.entity.aitool.AIPortToolMakerTemplateInstance;
 import ai.mcpdirect.backend.dao.mapper.account.TeamMapper;
 import org.apache.ibatis.annotations.Insert;
 import org.apache.ibatis.annotations.Param;
@@ -13,28 +12,32 @@ import java.util.List;
 public interface ToolMakerTemplateMapper {
     String TABLE = "aitool.tool_maker_template";
     String TABLE_JOIN = "aitool.tool_maker_template tmt";
+//    String SELECT_FIELDS = "id, created, status, type, name, agent_id agentId, last_updated lastUpdated," +
+//            "user_id userId,config,inputs";
     String SELECT_FIELDS = "id, created, status, type, name, agent_id agentId, last_updated lastUpdated," +
-            "user_id userId,config,inputs";
+            "user_id userId,inputs";
+//    String SELECT_FIELDS_JOIN = "tmt.id, tmt.created, tmt.status, tmt.type, tmt.name, tmt.agent_id agentId," +
+//            "tmt.last_updated lastUpdated, tmt.user_id userId,tmt.config,tmt.inputs";
     String SELECT_FIELDS_JOIN = "tmt.id, tmt.created, tmt.status, tmt.type, tmt.name, tmt.agent_id agentId," +
-            "tmt.last_updated lastUpdated, tmt.user_id userId,tmt.config,tmt.inputs";
-    String INSERT_FIELDS = "id, created, status, type, name, agent_id, last_updated, user_id,config,inputs";
-    
+            "tmt.last_updated lastUpdated, tmt.user_id userId,tmt.inputs";
+//    String INSERT_FIELDS = "id, created, status, type, name, agent_id, last_updated, user_id,config,inputs";
+    String INSERT_FIELDS = "id, created, status, type, name, agent_id, last_updated, user_id,inputs";
     @Insert("INSERT INTO "+TABLE+"("+INSERT_FIELDS+") VALUES\n" +
-            "(#{id}, #{created}, #{status}, #{type}, #{name}, #{agentId}, #{lastUpdated}, #{userId},#{config},#{inputs})")
+            "(#{id}, #{created}, #{status}, #{type}, #{name}, #{agentId}, #{lastUpdated}, #{userId},#{inputs})")
     int insertToolMakerTemplate(AIPortToolMakerTemplate template);
 
     @Update("<script>UPDATE "+TABLE + "\n" + """
             SET last_updated=#{lastUpdated}
             <if test="name!=null">,name=#{name}</if>
             <if test="status!=null">,status=#{status}</if>
-            <if test="config!=null">,config=#{config}</if>
             <if test="inputs!=null">,inputs=#{inputs}</if>
             WHERE id=#{id} AND user_id=#{userId}</script>""")
     int updateToolMakerTemplate(@Param("id")long id,
                                 @Param("userId")long userId,
                                 @Param("name")String name,
                                 @Param("status")Integer status,
-                                @Param("config")String config,@Param("inputs")String inputs,
+//                                @Param("config")String config,
+                                @Param("inputs")String inputs,
                                 @Param("lastUpdated")long lastUpdated);
 
     @Select("SELECT "+SELECT_FIELDS+" FROM "+TABLE+" WHERE id=#{id}")