+++ src/main/java/ai/mcpdirect/backend/dao/mapper/aitool/VirtualToolMapper.java
@@ -28,11 +28,11 @@ public interface VirtualToolMapper {
             "t.name,t.agent_id agentId," +
             "t.agent_status agentStatus FROM "+TABLE_JOIN_NAME+
             " LEFT JOIN "+AIToolMapper.TABLE_JOIN_NAME+" ON t.id=vt.tool_id\n"+
-            " WHERE vt.maker_id=#{makerId} AND vt.status>-1")
-    List<AIPortVirtualTool> selectVirtualToolByMakerId(@Param("makerId")long makerId);
+            " WHERE vt.maker_id=#{makerId} AND vt.status>-1 AND vt.last_updated>#{lastUpdated}")
+    List<AIPortVirtualTool> selectVirtualToolsByMakerId(@Param("makerId")long makerId, @Param("lastUpdated")long lastUpdated);
     @Select("SELECT "+SELECT_JOIN_FIELDS+",CONCAT(vt.tags,',',t.tags) tags," +
             "t.name,t.agent_id agentId,t.agent_status agentStatus FROM "+TABLE_JOIN_NAME+
             " LEFT JOIN "+AIToolMapper.TABLE_JOIN_NAME+" ON t.id=vt.tool_id\n"+
-            " WHERE vt.user_id=#{userId} AND vt.status>-1")
-    List<AIPortVirtualTool> selectVirtualTools(@Param("userId")long userId);
+            " WHERE vt.user_id=#{userId} AND vt.status>-1 AND vt.last_updated>#{lastUpdated}")
+    List<AIPortVirtualTool> selectVirtualTools(@Param("userId")long userId,@Param("lastUpdated")long lastUpdated);
 }